requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.24
langchain-pinecone==0.2.8
langchain-tests==0.3.20
langchain-text-splitters==0.3.8
langsmith==0.4.1
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.5.0
numpy==2.3.1
openai==1.90.0
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.3.0
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pinecone==7.2.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.1
pypdfium2==4.30.1
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==3.2.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.1
syrupy==4.9.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
vcrpy==7.0.0
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
